# relational-database-sql-tool/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the relational-database-sql-tool API Documentation.",
        "title": "Relational Database Sql Tool API",
        "contact": {
            "name": "Utopiops",
            "url": "http://www.utopiops.com",
            "email": "contactus@utopiops.com"
        },
        "license": {
            "name": "Copyright (C) Utopiops"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/rds/environment/name/{env_name}/database/name/{db_name}/list_table": {
            "post": {
                "description": "get list of all tables for a specific database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of user environment in our system",
                        "name": "env_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of user's database server",
                        "name": "db_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "username, password and database name are necessary to get list of tables",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rds.RDSQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "rds.RDSQuery": {
            "type": "object",
            "properties": {
                "database": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "statement": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}